requirements.txt

bs4
lxml
EbookLib
colorama
python-poppler
PyMuPDF
PyPDF2
pillow
tqdm
mobi

datasketch
spacy
ftfy
numpy
scipy
nltk
